requirements.txt
requests
pytest
pyyaml
loguru
allure-pytest
pytest-html
pyautogui
pyperclip
schedule